feat:make seprate types folder and all types defines in this folder and clean code
