--- src/screens/signUpScreen/type.ts
@@ -1,10 +1,0 @@
-export type CustomInputProps ={
-    placeholder?: string;
-    value: string;
-    onChangeText: (text: string) => void;
-    secureTextEntry?: boolean;
-    keyboardType?: 'default' | 'email-address' | 'numeric' | 'phone-pad';
-    email: string;
-    password: string;
-    name: string;
-  }
