+++ src/types/password.ts
@@ -1,0 +1,11 @@
+export interface UpdatePasswordPayload {
+    oldPassword: string;
+    newPassword: string;
+  }
+  
+  export interface PasswordUpdateState {
+    oldPassword: string;
+    newPassword: string;
+    confirmNewPassword: string;
+    isLoading: boolean;
+  }
