+++ src/types/signup.ts
@@ -1,0 +1,12 @@
+export interface SignUpState {
+    name: string;
+    setName: (name: string) => void;
+    email: string;
+    setEmail: (email: string) => void;
+    password: string;
+    setPassword: (password: string) => void;
+    loading: boolean;
+    showError: boolean;
+    emailError: string | null;
+    handleRegister: () => Promise<{ name: string; email: string; password: string } | undefined>;
+  }
